Add dynamic import wrapper for `landingPage.js`. Bump patch version...
This patch bump is a fix for JS errors on KCC landing-pages that don't
have a target X form.

- The module `landingPage.js` requires an element to match
`document.querySelector('a[href="#page-top"')` in the page.
- However, some landing pages do not have the targetX form sections which
is the reason for the modules' JS.
- This caused errors to be thrown from the `landinPage.js` module.

To fix this, the dynamic import of the module `landinPage.js` was wrapped in a check
for the required elements.